+++ test/support/helpers.ex
@@ -18,29 +18,16 @@ defmodule ExBuffer.Helpers do
   @doc false
   @spec start_ex_buffer(keyword()) :: {:ok, GenServer.name()} | {:error, ExBuffer.error()}
   def start_ex_buffer(opts \\ []) do
-    opts =
-      opts
-      |> Keyword.put_new(:name, :ex_buffer)
-      |> Keyword.put_new(:flush_callback, flush_callback(:ex_buffer))
-
-    case start_supervised({ExBuffer, opts}) do
-      {:ok, _} -> {:ok, Keyword.get(opts, :name)}
-      {:error, {reason, _}} -> {:error, reason}
-    end
+    name = Keyword.get(opts, :name, ExBuffer)
+    opts = Keyword.put_new(opts, :flush_callback, flush_callback(name))
+    start_buffer({ExBuffer, opts})
   end
 
   @doc false
   @spec start_test_buffer(keyword()) :: {:ok, GenServer.name()} | {:error, ExBuffer.error()}
   def start_test_buffer(opts \\ []) do
-    opts =
-      opts
-      |> Keyword.put_new(:name, :test_buffer)
-      |> Keyword.put(:flush_meta, self())
-
-    case start_supervised({ExBuffer.TestBuffer, opts}) do
-      {:ok, _} -> {:ok, Keyword.get(opts, :name)}
-      {:error, {reason, _}} -> {:error, reason}
-    end
+    opts = Keyword.put(opts, :flush_meta, self())
+    start_buffer({ExBuffer.TestBuffer, opts})
   end
 
   ################################
@@ -50,5 +37,22 @@ defmodule ExBuffer.Helpers do
   defp flush_callback(name) do
     destination = self()
     fn data, opts -> send(destination, {name, data, opts}) end
+  end
+
+  defp start_buffer(child_spec) do
+    case start_supervised(child_spec) do
+      {:ok, pid} -> get_buffer_name(pid)
+      {:error, {{_, {_, _, reason}}, _}} -> {:error, reason}
+      {:error, {reason, _}} -> {:error, reason}
+    end
+  end
+
+  defp get_buffer_name(pid) do
+    pid
+    |> Process.info()
+    |> case do
+      nil -> {:error, :not_found}
+      info -> {:ok, Keyword.get(info, :registered_name)}
+    end
   end
 end